# Smart string hashing