Use full string hashing when a lot of collisions are generated.
- detect when collision chain is too long.
  "average maximum" chain of table with fill-factor 1.0 is near 7, so if collision chain is longer than 10, then it is bad sign.
- calculate "full" hash for new strings in long collision chain.
- use "bloom" filter to bookkeeping existence of strings with "full" hash.
- refill "bloom" on string sweeping.

Separately:
- first commit adds checking for hash equality when collision chain is traversed.
- last commit adds optional usage of strong hash function for "full hash".

Obsoletes #171 and #169, closes #168