@funny-falcon 2016-05-04
Benchmark: https://gist.github.com/funny-falcon/685dbfaea16b5919e6c84ab1b156d2f6