@funny-falcon 2016-05-06
Copy benchmark numbers:
Benchmarks:
1. small safe strings: strings are short, and fully covered by sparse hash
2. safe strings: 136bytes, and differs by last bytes (which covered by sparse hash)
3. long half-safe strings: 264bytes, diff is in a middle, sparse hash covers less frequently changed bytes of a diff
4. long unsafe strings: 392bytes, diff is not covered by sparse hash, diff is in first third.
5. long unsafe strings: 392bytes, diff is not covered by sparse hash, diff is in second third
6. long safe string: 392bytes, diff is at the end, so sparse hash covers much changed bytes.

Bench | Count | Unpatched | Patched | PatchedStrong
---------|---------|---------------|------------|-------------------
1 | 1000000 | 0.19s | 0.19s | 0.19s 
2 | 500000 | 0.16s | 0.15s | 0.15s
3 | 100000 | 0.66s | 0.11s | 0.13s
4 | 10000   | 2.99s | 0.012s | 0.016s
5 | 10000   | 5.12s | 0.016s | 0.024s
6 | 300000 | 0.16s | 0.14s | 0.14s