@bungle 2016-05-09
This fixes issues in my benchmark¹ as well, and I cannot see any side effects either. Without the patch and running something like 10 000 iterations it totally kills unpatched LuaJIT.

¹⁾ https://github.com/bungle/lua-resty-template/blob/master/lib/resty/template/microbenchmark.lua