@funny-falcon 2016-06-27
Why "unly for *nix"? It is cause of `/dev/urandom` ?
I just don't know how to easily get secure random within Windows :-(
If you know how to, please help me.