@kirbyfan64 2016-06-27
@funny-falcon https://msdn.microsoft.com/en-us/library/windows/desktop/aa379942(v=vs.85).aspx