@funny-falcon 2016-06-27
Could you try to compile on Windows now with `LUAJIT_SMART_STRING` set to 2 ?