@data-man 2016-06-27
Sorry , I'm writing from the phone.
And this slow (but strong) mode is necessary?

As an experiment, try xxHash32 (xxHash64 for x64). It's really fast.
https://github.com/Cyan4973/xxHash