@funny-falcon 2016-06-27
Dmitry, I've played a lot with hash functions, and I know about xxhash.

xxhash32 is not more secure and not much faster than my fast hash (maybe
even slower, i don't remember exact numbers) (LUAJIT_SMART_STRING=1) . But
it is much more complex.

Secure mode has to exist for those, who wants more security. It is compile
time option and it is not default. So if you doesn't want it, you will not
pay for.