@funny-falcon 2016-06-27
@data-man , I'm really sure, that "secure" hash could at least twice faster (by mixing 2blocks-8bytes per siphash round instead of 1block-4bytes). It still will be safe for this use-case.

But I don't have much authority to people to believe such claim :-(