@funny-falcon 2016-06-27
@bungle can you rerun tests, please? Both with SMART_STRING=1 and 2 (and unpatched).

I've noticed, that previous version were a bit slower for uncontendent case. So I've change heuristic a bit. I'd like to see, if there is a difference in realworld case.