@bungle 2016-06-27
SMART_STRING = 1:

```
luajit -e 'require "resty.template.microbenchmark".run(10000)'

Running 10000 iterations in each test
    Parsing Time: 0.041476
Compilation Time: 0.269675 (template)
Compilation Time: 0.000118 (template, cached)
  Execution Time: 0.354958 (same template)
  Execution Time: 0.036355 (same template, cached)
  Execution Time: 0.438314 (different template)
  Execution Time: 0.074922 (different template, cached)
  Execution Time: 0.426644 (different template, different context)
  Execution Time: 0.065288 (different template, different context, cached)
      Total Time: 1.707750

```

SMART_STRING = 2:

```
luajit -e 'require "resty.template.microbenchmark".run(10000)'

Running 10000 iterations in each test
    Parsing Time: 0.043237
Compilation Time: 0.265141 (template)
Compilation Time: 0.000123 (template, cached)
  Execution Time: 0.356360 (same template)
  Execution Time: 0.037365 (same template, cached)
  Execution Time: 0.453068 (different template)
  Execution Time: 0.068437 (different template, cached)
  Execution Time: 0.428845 (different template, different context)
  Execution Time: 0.071860 (different template, different context, cached)
      Total Time: 1.724436
```

SMART_STRING = 0:

```
luajit -e 'require "resty.template.microbenchmark".run(10000)'

Running 10000 iterations in each test
    Parsing Time: 0.042821
Compilation Time: 0.262019 (template)
Compilation Time: 0.000114 (template, cached)
  Execution Time: 0.364316 (same template)
  Execution Time: 0.038646 (same template, cached)
  Execution Time: 5.617718 (different template)
  Execution Time: 1.313856 (different template, cached)
  Execution Time: 5.867694 (different template, different context)
  Execution Time: 1.365527 (different template, different context, cached)
      Total Time: 14.872711
```

I'm not sure if this is a "real world" test, but it shows a potential problem that is easy to trigger even with a lot smaller iteration counts than the above 10,000. Both SMART_STRING = 1 and SMART_STRING = 2 work about equally the same (probably somewhat in a error margin), but SMART_STRING = 0 renders exponentially worse results (SMART_STRING = 1 and SMART_STRING = 2 behave linerially when the iteration count is raised). So, I still hope these would be merged.