@funny-falcon 2016-06-28
@bungle , thank you. Now "template" and "same template" lines look like not sufferring from patch.